+++ src/mc/dsp/widget/TempoDetect.cpp
@@ -40,7 +40,7 @@ auto TempoDetect::operator()(Span<float> input, float sampleRate) -> float
 {
 
     auto const levels        = wt_.levels();
-    auto const maxDecimation = std::pow(2.0F, levels - 1);
+    auto const maxDecimation = std::pow(2.0F, static_cast<float>(levels - 1));
     auto const minNdx        = std::floor(60.0F / 220.0F * (sampleRate / maxDecimation));
     auto const maxNdx        = std::floor(60.0F / 40.0F * (sampleRate / maxDecimation));
